src/services/image_segmentation.py
"""
图像分割服务
Image Segmentation Service

基于 SAM (Segment Anything Model) 的图像分割功能
支持：
- 自动分割所有元素
- 文字识别和分离
- 主体提取
- 背景分离
- Inpainting 修复
"""

import os
import base64
import logging
from typing import Optional, List, Dict, Any, Tuple
from io import BytesIO
from PIL import Image
import numpy as np

logger = logging.getLogger(__name__)


class ImageSegmentationService:
    """
    图像分割服务
    
    提供基于 AI 的图像分割功能：
    1. SAM 语义分割
    2. OCR 文字识别
    3. Inpainting 背景修复
    """
    
    def __init__(self):
        """初始化分割服务"""
        self._sam_model = None
        self._ocr_model = None
        self._inpaint_model = None
    
    async def segment_image(
        self, 
        image_base64: str,
        extract_text: bool = True,
        extract_subjects: bool = True,
        extract_background: bool = True
    ) -> Dict[str, Any]:
        """
        分割图像为多个图层
        
        Args:
            image_base64: 图像的 base64 编码
            extract_text: 是否提取文字
            extract_subjects: 是否提取主体
            extract_background: 是否提取背景
            
        Returns:
            {
                "layers": [...],  # 图层列表
                "text_regions": [...],  # 文字区域
                "original_width": int,
                "original_height": int
            }
        """
        try:
            # 解析图像
            image = self._decode_base64_image(image_base64)
            width, height = image.size
            
            layers = []
            text_regions = []
            
            # 1. 提取文字
            if extract_text:
                text_regions = await self._extract_text(image)
            
            # 2. 分割主体
            if extract_subjects:
                subject_layers = await self._segment_subjects(image)
                layers.extend(subject_layers)
            
            # 3. 提取背景
            if extract_background:
                background_layer = await self._extract_background(image, layers)
                if background_layer:
                    layers.insert(0, background_layer)
            
            return {
                "layers": layers,
                "text_regions": text_regions,
                "original_width": width,
                "original_height": height
            }
            
        except Exception as e:
            logger.error(f"图像分割错误: {e}")
            raise
    
    async def _extract_text(self, image: Image.Image) -> List[Dict[str, Any]]:
        """
        提取图像中的文字
        
        使用 OCR 识别文字位置和内容
        
        TODO: 集成真实 OCR 模型 (PaddleOCR / EasyOCR / Tesseract)
        """
        try:
            # 模拟 OCR 结果
            # 实际实现需要：
            # 1. 调用 OCR API 或本地模型
            # 2. 解析返回的文字框位置
            # 3. 估计字体大小和颜色
            
            text_regions = []
            
            # 占位实现 - 返回空列表
            # 后续集成 PaddleOCR:
            # from paddleocr import PaddleOCR
            # ocr = PaddleOCR(use_angle_cls=True, lang='ch')
            # result = ocr.ocr(np.array(image), cls=True)
            
            return text_regions
            
        except Exception as e:
            logger.error(f"文字提取错误: {e}")
            return []
    
    async def _segment_subjects(self, image: Image.Image) -> List[Dict[str, Any]]:
        """
        分割图像中的主体
        
        使用 SAM 自动分割所有独立元素
        
        TODO: 集成 SAM 2.1 或 Grounded SAM
        """
        try:
            # 模拟分割结果
            # 实际实现需要：
            # 1. 加载 SAM 模型
            # 2. 运行自动分割
            # 3. 提取每个分割区域的蒙版
            
            layers = []
            
            # 占位实现
            # 后续集成 SAM:
            # from segment_anything import sam_model_registry, SamAutomaticMaskGenerator
            # sam = sam_model_registry["vit_h"](checkpoint="sam_vit_h_4b8939.pth")
            # mask_generator = SamAutomaticMaskGenerator(sam)
            # masks = mask_generator.generate(np.array(image))
            
            return layers
            
        except Exception as e:
            logger.error(f"主体分割错误: {e}")
            return []
    
    async def _extract_background(
        self, 
        image: Image.Image, 
        foreground_layers: List[Dict[str, Any]]
    ) -> Optional[Dict[str, Any]]:
        """
        提取背景层
        
        通过移除前景元素获取纯背景
        """
        try:
            width, height = image.size
            
            # 创建背景图层
            background_layer = {
                "id": "background",
                "type": "background",
                "mask_base64": "",
                "content_base64": None,
                "bbox": [0, 0, width, height],
                "metadata": {"description": "背景图层"}
            }
            
            return background_layer
            
        except Exception as e:
            logger.error(f"背景提取错误: {e}")
            return None
    
    async def remove_element_and_inpaint(
        self, 
        image_base64: str,
        mask_base64: str
    ) -> Dict[str, Any]:
        """
        移除元素并修复背景
        
        使用 Inpainting 技术填充被移除区域
        
        Args:
            image_base64: 原始图像
            mask_base64: 要移除区域的蒙版
            
        Returns:
            {"result_base64": str, "width": int, "height": int}
        """
        try:
            # 解析图像和蒙版
            image = self._decode_base64_image(image_base64)
            mask = self._decode_base64_image(mask_base64)
            
            # TODO: 实现 Inpainting
            # 可选方案：
            # 1. Stable Diffusion Inpainting
            # 2. LaMa Inpainting
            # 3. OpenCV Inpainting (基础)
            
            # 占位实现 - 返回原图
            result_base64 = self._encode_image_base64(image)
            
            return {
                "result_base64": result_base64,
                "width": image.size[0],
                "height": image.size[1]
            }
            
        except Exception as e:
            logger.error(f"Inpainting 错误: {e}")
            raise
    
    async def edit_text_in_image(
        self,
        image_base64: str,
        text_edits: List[Dict[str, Any]]
    ) -> Dict[str, Any]:
        """
        编辑图像中的文字
        
        工作流：
        1. 定位原始文字
        2. 使用 Inpainting 移除原始文字
        3. 渲染新文字
        4. 合成最终图像
        
        Args:
            image_base64: 原始图像
            text_edits: 编辑操作列表
                [{"region_id": "...", "new_text": "...", "font_size": 24, "color": "#fff"}]
                
        Returns:
            {"result_base64": str, "width": int, "height": int}
        """
        try:
            image = self._decode_base64_image(image_base64)
            
            # TODO: 实现完整的文字编辑工作流
            # 1. 根据 region_id 找到文字区域
            # 2. 创建文字区域蒙版
            # 3. 使用 Inpainting 移除原始文字
            # 4. 使用 PIL/Pillow 渲染新文字
            # 5. 合成结果
            
            # 占位实现
            result_base64 = self._encode_image_base64(image)
            
            return {
                "result_base64": result_base64,
                "width": image.size[0],
                "height": image.size[1]
            }
            
        except Exception as e:
            logger.error(f"文字编辑错误: {e}")
            raise
    
    def _decode_base64_image(self, image_base64: str) -> Image.Image:
        """解码 base64 图像"""
        # 处理 data URL
        if "," in image_base64:
            image_base64 = image_base64.split(",")[1]
        
        image_data = base64.b64decode(image_base64)
        return Image.open(BytesIO(image_data))
    
    def _encode_image_base64(self, image: Image.Image, format: str = "PNG") -> str:
        """编码图像为 base64"""
        buffer = BytesIO()
        image.save(buffer, format=format)
        return base64.b64encode(buffer.getvalue()).decode()


# 服务单例
_segmentation_service: Optional[ImageSegmentationService] = None


def get_segmentation_service() -> ImageSegmentationService:
    """获取分割服务实例"""
    global _segmentation_service
    if _segmentation_service is None:
        _segmentation_service = ImageSegmentationService()
    return _segmentation_service
