requirements.txt
# Core Dependencies
anthropic>=0.18.0          # Claude API
openai>=1.12.0             # OpenAI API / ALLAPI兼容
pydantic>=2.5.0            # 数据验证
python-dotenv>=1.0.0       # 环境变量

# API Framework
fastapi>=0.109.0           # Web框架
uvicorn[standard]>=0.27.0  # ASGI服务器
python-multipart>=0.0.6    # 文件上传
websockets>=12.0           # WebSocket支持

# Execution & Tools
docker>=7.0.0              # Docker容器管理
redis>=5.0.0               # 缓存和任务队列
httpx>=0.26.0              # HTTP客户端
beautifulsoup4>=4.12.0     # HTML解析
lxml>=5.0.0                # XML解析
modelcontextprotocol>=1.0.1  # 官方 MCP 客户端

# Browser Automation
playwright>=1.40.0         # 浏览器自动化

# Data Processing
pandas>=2.1.0              # 数据分析
numpy>=1.26.0              # 数值计算

# Database
sqlalchemy>=2.0.0          # ORM
alembic>=1.13.0            # 数据库迁移

# Testing
pytest>=7.4.0              # 测试框架
pytest-asyncio>=0.23.0     # 异步测试
pytest-cov>=4.1.0          # 测试覆盖率
pytest-mock>=3.12.0        # Mock工具

# Development
black>=24.0.0              # 代码格式化
ruff>=0.1.0                # Linter
mypy>=1.8.0                # 类型检查
pre-commit>=3.6.0          # Git hooks

# Logging & Monitoring
loguru>=0.7.0              # 日志
structlog>=24.1.0          # 结构化日志

# Utils
tenacity>=8.2.0            # 重试机制
click>=8.1.0               # CLI工具
pyyaml>=6.0.0              # YAML配置
jsonschema>=4.20.0         # JSON验证
aiofiles>=23.0.0           # 异步文件操作

# Document & Media (Optional)
markdown>=3.5.0            # Markdown处理
python-pptx>=0.6.21        # PPT生成
